tts_service/requirements.txt
coqui-tts
torch
tqdm
spacy
pydub